+++ packages/core/styles/src/index.ts
@@ -6,6 +6,11 @@
 
 export * from "./components";
 export { vars } from "./utils";
+export {
+  gradients,
+  createLinearGradient,
+  createBorderGradient,
+} from "./gradients";
 
 export { badge } from "./packages/atomic/badge";
 export { box } from "./packages/atomic/box";
@@ -33,6 +38,7 @@ export { link } from "./packages/atomic/link";
 export { tooltip } from "./packages/atomic/tooltip";
 export { thumbnail } from "./packages/atomic/thumbnail";
 export { textarea } from "./packages/atomic/textarea";
+export { progressBar } from "./packages/atomic/progressBar";
 
 export { accordion } from "./packages/composite/accordion";
 export { alert } from "./packages/composite/alert";
@@ -50,7 +56,7 @@ export { scrollPane } from "./packages/composite/scroll-pane";
 export { calendar } from "./packages/patterns/calendar";
 export { editor } from "./packages/patterns/editor";
 
-
+/* Types */
 export type { BoxSprinkle } from "./packages/atomic/box";
 export type { IconButtonSprinkle } from "./packages/atomic/iconButton";
 export type { TextSprinkle } from "./packages/atomic/text";