+++ packages/react/src/atomic/Icon/src/index.ts
@@ -2,4 +2,10 @@ import { Icon } from "./Icon";
 
 export { Icon } from "./Icon";
 export type { IconProps } from "./icon.types";
+export {
+  isGradient,
+  canUseGradient,
+  isValidColorValue,
+  GRADIENT_COLORS,
+} from "./utils/gradient";
 export default Icon;