+++ packages/react/src/atomic/Icon/src/utils/gradient/gradientFactory/index.ts
@@ -1,0 +1,1 @@
+export { getGradientStops, createSvgGradientDef } from "./gradientFactory";