+++ packages/react/src/atomic/Icon/src/utils/gradient/index.ts
@@ -1,0 +1,7 @@
+export { applyGradientToSvg } from "./gradientProcessor";
+export {
+  isGradient,
+  canUseGradient,
+  isValidColorValue,
+  GRADIENT_COLORS,
+} from "./colorValidation";