+++ packages/react/src/atomic/Icon/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 Icons are used to visually communicate core parts of the product and available actions. They can act as wayfinding tools to help merchants more easily understand where they are in the product, and common interaction patterns that are available.
 
+## 2025-08-01 `3.2.1`
+
+### 🐛 Bug fixes
+
+- Rename `ai-interactive` to `ai-generative` in `Icon` component. ([#327](https://github.com/TiendaNube/nimbus-design-system/pull/327) by [@joacotornello](https://github.com/joacotornello))
+
+## 2025-07-30 `3.2.0`
+
+### 🎉 New features
+
+- Added `ai-interactive` appearance to `Icon` component. ([#321](https://github.com/TiendaNube/nimbus-design-system/pull/321) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-06-27 `3.1.0`
 
 ### 🎉 New features